+++ src/app/shared/directives/formatted.directive.spec.ts
@@ -1,0 +1,8 @@
+import { FormattedDirective } from './formatted.directive';
+
+describe('FormattedDirective', () => {
+  it('should create an instance', () => {
+    const directive = new FormattedDirective();
+    expect(directive).toBeTruthy();
+  });
+});